# SpriteBatch optimisation: Only calculate projection matrix when viewport changes