I was inspired by PR #5456 by @nkast and came up with another related optimisation :)

This PR only re-calculates the projection matrix when the viewport changes. Given that the viewport will normally rarely change, this results in a not bad performance boost (takes about 16% less time). It's not quite as good a boost as PR #5456, but decent nonetheless :)

Benchmarks (using BenchmarkDotNet):
```c#
spriteBatch.Begin();
spriteBatch.End();
```
Benchmark | Median time | Scale<br>(Original baseline) | Scale<br>(PR #5456 baseline)
:--- | ---:|:---:|:---:
Original code (before PR #5456) | 383.2844 ns | 1.00 | 1.28
After PR #5456 | 299.4846 ns | 0.78 | 1.00
This PR | 253.0242 ns | 0.66 | **0.84**