@nkast 2017-02-05
Good one @MichaelDePiazzi !

We can also unroll the `CreateOrthographicOffCenter`. Because half of the inputs is 0, some divisions will always equal 1 or -1. 
BWT ,  Do we need all those doubles in `CreateOrthographicOffCenter()`?