@MichaelDePiazzi 2017-02-05
@nkast I'm not sure if there'll be much advantage to unrolling `CreateOrthographicOffCenter` as it's now called rarely anyway (only when the viewport size changes). There may be some advantage to unrolling the `Matrix.Multiply` though. I will look into them both and see what I can do in regards to unrolling. Thanks! :)