@nkast 2017-02-05
Yes , you are right! the CreateOrthographicOffCenter is almost never called now. And actually the cases of calling Begin/End frequently are very rare. But #5456 was a simple change so why not.