@MichaelDePiazzi 2017-02-05
@nkast Well, I did successfully unroll both the operations and it was a bit faster. `CreateOrthographicOffCenter` went from ~35ns to ~2ns, and `Matrix.Multiply` went from ~69ns to ~22ns. But `CreateOrthographicOffCenter` is now rarely called, so that one may not be worthwhile as it will make it more difficult to maintain.

But I think it might be best to submit this in a separate PR as it will make the code more complex and more difficult to review (so less likely to get merged quickly). And if you're saying there's some uncertainty with the Z values, then it's probably best to ensure that it's right before we unroll it.