@tomspilman 2017-02-07
> And actually the cases of calling Begin/End frequently are very rare. 

Actually I've seen plenty of games in porting that call `Begin/End` a lot more than they should.  This should be a great optimization for those titles.

> I'm not sure if there'll be much advantage to unrolling CreateOrthographicOffCenter 

I agree...  you would have to have some crazy game code that constantly changes viewport size during a frame for that optimization to matter.  I've yet to see a game ever do this... at most it is a couple of viewport changes a frame for UI stuff.

So is this ready to merge @MichaelDePiazzi @nkast ?