@MichaelDePiazzi 2017-02-07
@tomspilman I've just pushed 1 more commit which simplifies the matrix multiplication (as @nkast originally suggested). In my benchmarks I actually found it to be a tiny bit faster, but the main benefit is that it simplifies the code. So after this builds, it should be ready to merge.